+++ test/network/network.cpp
@@ -1,13 +1,13 @@
 #include <gtest/gtest.h>
 
 extern "C" {
-    #include "network/network.h"
+#include "network/network.h"
 }
 
 #include <math.h>
 
 float sigmoid(float val, int is_deravative) {
-    float result = 1.0f / (1.0f + exp(-val));
+    float result = 1.0f / (1.0f + expf(-val));
     if (is_deravative == 1) {
         return result * (1.0f - result);
     }
@@ -27,8 +27,8 @@ TEST(NetworkTest, RandomFloat) {
 }
 
 TEST(NetworkTest, MatrixMultiply) {
-    float *a = (float*)malloc(sizeof(float) * 3);
-    float *b = (float*)malloc(sizeof(float) * 3);
+    float *a = (float *)malloc(sizeof(float) * 3);
+    float *b = (float *)malloc(sizeof(float) * 3);
 
     a[0] = 1.0f;
     a[1] = 2.0f;
@@ -37,7 +37,7 @@ TEST(NetworkTest, MatrixMultiply) {
     b[1] = 5.0f;
     b[2] = 6.0f;
 
-    float *c = (float*)malloc(sizeof(float) * 9);
+    float *c = (float *)malloc(sizeof(float) * 9);
     matrix_multiply(a, b, c, 3, 1, 3);
     ASSERT_FLOAT_EQ(c[0], 4.0f);
     ASSERT_FLOAT_EQ(c[2], 12.0f);
@@ -59,7 +59,7 @@ TEST(NetworkTest, GetLayer) {
     ASSERT_NE(layer->weights, nullptr);
     ASSERT_NE(layer->bias, nullptr);
     ASSERT_NE(layer->output, nullptr);
-    
+
     ASSERT_EQ(layer->length, layer_length);
 
     free_layer(layer);
@@ -67,7 +67,7 @@ TEST(NetworkTest, GetLayer) {
 
 TEST(NetworkTest, GetNeuralNetwork) {
     size_t layer_length = 3;
-    size_t *layer_lengths = (size_t*)malloc(sizeof(size_t) * layer_length);
+    size_t *layer_lengths = (size_t *)malloc(sizeof(size_t) * layer_length);
     layer_lengths[0] = 2;
     layer_lengths[1] = 3;
     layer_lengths[2] = 4;
@@ -92,7 +92,7 @@ TEST(NetworkTest, GetNeuralNetwork) {
 
 TEST(NetworkTest, FreeDataset) {
     size_t layer_length = 1;
-    size_t *layer_lengths = (size_t*)malloc(sizeof(size_t) * layer_length);
+    size_t *layer_lengths = (size_t *)malloc(sizeof(size_t) * layer_length);
     layer_lengths[0] = 2;
     neural_network_t *nn = get_neural_network(layer_length, layer_lengths, 2, nullptr);
 
@@ -112,11 +112,11 @@ TEST(NetworkTest, FreeLayer) {
 TEST(NetworkTest, ComputeNetwork) {
     size_t layer_length = 1;
     size_t inputs_length = 1;
-    size_t *layer_lengths = (size_t*)malloc(sizeof(size_t) * layer_length);
+    size_t *layer_lengths = (size_t *)malloc(sizeof(size_t) * layer_length);
     layer_lengths[0] = 1;
     neural_network_t *nn = get_neural_network(layer_length, layer_lengths, inputs_length, &sigmoid);
 
-    float *inputs = (float*)malloc(sizeof(float) * inputs_length);
+    float *inputs = (float *)malloc(sizeof(float) * inputs_length);
     inputs[0] = 0.2f;
 
     nn->layers[0]->weights[0] = 0.5f;
